requirements.txt
coverage==5.1
#pudb==2019.2  # python debugger
evdev==1.3.0
time
rpi.GPIO==0.7.0